Move CompilerResources to public API
This type is used in BaseRule so it should be part of the public API.
