+++ detekt-api/src/main/kotlin/io/gitlab/arturbosch/detekt/api/CompilerResources.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.api.internal
+package io.gitlab.arturbosch.detekt.api
 
 import org.jetbrains.kotlin.config.LanguageVersionSettings
 import org.jetbrains.kotlin.resolve.calls.smartcasts.DataFlowValueFactory
